# openapi.yaml
openapi: 3.1.0
info:
  title: OssPilot Tenant API
  version: 0.1.0
paths:
...
  /api/versions/{version_id}:
    get:
      summary: Version metadata
      responses:
        "200":
          description: summary
        "404":
          description: not found
    delete:
      summary: Delete an archived version
      responses:
        "200":
          description: deleted
...